+++ app/controllers/staff/users_controller.rb
@@ -70,7 +70,7 @@ class Staff::UsersController < Staff::BaseController
   end
 
   def user_params
-    params.require(:user).permit(:name, :email, :role, :organisation_id)
+    params.require(:user).permit(:name, :email, :organisation_id)
   end
 
   def id