+++ app/models/user.rb
@@ -6,21 +6,11 @@ class User < ApplicationRecord
   validates :email, format: {with: URI::MailTo::EMAIL_REGEXP}
   validates :email, with: :email_cannot_be_changed_after_create, on: :update
 
-  enum role: {
-    administrator: "administrator",
-  }
-
-  attribute :role, :string, default: "administrator"
-
   FORM_FIELD_TRANSLATIONS = {
     organisation_id: :organisation,
   }.freeze
 
   scope :active, -> { where(active: true) }
-
-  def role_name
-    I18n.t("activerecord.attributes.user.roles.#{role}")
-  end
 
   delegate :service_owner?, :delivery_partner?, to: :organisation